+++ Library/CloudVOffice.Services/HR/Master/IDepartmentService.cs
@@ -13,9 +13,9 @@ namespace CloudVOffice.Services.HR.Master
     {
         public MessageEnum CreateDepartment(DepartmentDTO departmentDTO);
         public List<Department> GetDepartmentList();
-        public Department GetDepartmentById(Int64 departmentId);
+        public Department GetDepartmentById(int departmentId);
         public MessageEnum DepartmentUpdate(DepartmentDTO departmentDTO);
-        public MessageEnum DepartmentDelete(Int64 deprtmentid, Int64 DeletedBy);
+        public MessageEnum DepartmentDelete(int deprtmentid, Int64 DeletedBy);
         public List<Department> GetAllDepartmentGroups(); 
     }
 }
